+++ src/pelmeni/graphics/Frame.hpp
@@ -5,11 +5,11 @@
 
 namespace p2d { namespace graphics {
     struct Frame {
-        Frame(const sf::Time& t, const sf::Rect<float>& rect)
+        Frame(const sf::Time& t, const sf::Rect<int>& rect)
         : duration(t), frameRect(rect) {;}
         
         sf::Time duration;
-        sf::Rect<float> frameRect;
+        sf::Rect<int> frameRect;
     }; // struct Frame
 } // namespace graphics
 } // namespace p2d
